requirements.txt
transformers
pillow
streamlit
requests
fastapi
torch
uvicorn
gunicorn